chore: move language selector to separate component
