+++ frontend/src/layouts/AdminLayout.vue
@@ -6,19 +6,7 @@
 
         <q-toolbar-title> HOSM Nepal – {{ t('appTitle') }} </q-toolbar-title>
 
-        <q-select
-          v-model="selectedLang"
-          :options="langOptions"
-          option-value="value"
-          option-label="label"
-          emit-value
-          map-options
-          outlined
-          dense
-          class="lang-select q-ml-md"
-          popup-content-class="lang-select-options"
-          @update:model-value="changeLocale"
-        />
+        <LanguageSelector />
 
         <q-btn
           color="grey-8"
@@ -40,30 +28,12 @@
 </template>
 
 <script setup lang="ts">
-import { ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useAuthStore } from 'stores/auth';
+import LanguageSelector from 'components/LanguageSelector.vue';
 
-const { t, locale } = useI18n();
+const { t } = useI18n();
 const authStore = useAuthStore();
-
-const langOptions = [
-  { label: 'EN', value: 'en-US' },
-  { label: 'FR', value: 'fr' },
-];
-
-const selectedLang = ref(locale.value);
-
-watch(locale, (newLocale) => {
-  selectedLang.value = newLocale;
-});
-
-function changeLocale(newLocale: string) {
-  locale.value = newLocale as 'en-US' | 'fr';
-  if (typeof window !== 'undefined') {
-    localStorage.setItem('app-locale', newLocale);
-  }
-}
 
 function logout() {
   authStore.logout();
